Merged PR 148: save all colors of a product on saving into locker room
As discussed, we are storing all the colors, whether changed or not, to ensure the products stored in the locker room remain the same if the design files are changed the color.

https://www.screencast.com/t/xN3TU7FLVcok

as you can see in the above video I am only changing the color of base to United Yellow but after storing you can see we have stored all different parts colors.
![image.png](https://dev.azure.com/Custimoo/f28161ec-6c1b-486c-a623-ae20c4419714/_apis/git/repositories/60fa9341-c503-4982-94ca-4516dc4925a5/pullRequests/148/attachments/image.png)

Related work items: #336