Merged PR 150: Setup sceript new version is completed
here are the updates in the setup script
The new script file is set "setup_ubuntu2004_custimoo_builder_environment_v1.sh".
First of all, we have integrated the help command which can show you all the different parameter definitions that are used for running the setup script.

./setup_ubuntu2004_custimoo_builder_environment_v1.sh -h

ImageAs you can see in the above image we have -b or --builddirectoryname that can be used to define build direct where we can put our builds. by default the build directory is artifacts. FYI this artifacts directory will be used to deploy on staging and production

./setup_ubuntu2004_custimoo_builder_environment_v1.sh -b artifacts

the second parameter is -d --domain. the default domain is custimoo-builder.test

./setup_ubuntu2004_custimoo_builder_environment_v1.sh -b custimoo_builder.test

the third parameter is -u --apiurl which is used to create the API connection. The default value is http://custimoo-v2-backend.test which is used in the backend as a domain, but you can change it according to your requirements. Please note you can not overwrite .env.production and .env.staging files with this parameter as these files are used to deploy for production and staging.

./setup_ubuntu2004_custimoo_builder_environment_v1.sh -u http://custimoo-backend.test

the fourth parameter is -m --modes this argument is used to run build according to different environments. there are four different ones have development, production staging, and serve. by default mode is development that is used to test the build. serve mode is used to run locally for development, staging and production builds are used to release on the server. you can also pass multiple comma-separated values to run multiple builds for different environments.

./setup_ubuntu2004_custimoo_builder_environment_v1.sh -m development,staging

the last parameter is -t --buildtypes this parameter is used to run different types of builds. As we have implemented our customizer for two different types of websites one is with e-commerce(woo-commerce/Shopify) and other are ones that use our cart flow. for that, we have three different types of builds ecommercecustomizer, selfcustomizer, orderdetail. so you can specify which particular build you need to run with this parameter. by default build type is selfcustomizer. but you can also pass multiple comma-separated values to create multiple builds.

./setup_ubuntu2004_custimoo_builder_environment_v1.sh -t ecommercecustomizer,selfcustomizer,orderdetail

Related work items: #217