Merged PR 438: Merged PR 437: safe zone moved to back side for the text is fixed
Merged PR 437: safe zone moved to back side for the text is fixed

In case of text, the product safe zone of the front design is applied to the back side of the design. In the picture, there is a dot shown on number 11 which is a safe zone for the front Hummel logo.
Image

Related work items: #754, #758

Related work items: #754, #758