Merged PR 1077: Merged PR 1076: There are some different conditions set for self type compani...
Merged PR 1076: There are some different conditions set for self type companies which is why...

There are some different conditions set for self type companies which is why it is not shown. Also, on every Add to Cart button, the condition to show the button differs. So create a common component for Add to Cart and use the same component everywhere.

Related work items: #1369

Related work items: #1368, #1369